requirements.txt
customtkinter>=5.2.2
ffmpeg-python>=0.2.0
audio-separator[cpu]>=0.30.1
